Agregando transaction por tipo de moneda

+++ src/main/kotlin/com/ucb/edu/abc/mscompany/dto/request/TransactionDto.kt
@@ -31,7 +31,7 @@ data class TransactionDto(
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class TransactionAccountDto (
-        var accountId: Long,
+        var accountId: Int,
         var entityId: Int?,
         var auxiliaryId: Int?,
         var amountDebit: BigDecimal,
